add custom popperStyle to Select component
